Implemented blog card component
